Add files via upload
Pawsitive – Treat The Stray is a comprehensive platform designed to improve the welfare of stray animals. It offers emergency ambulance services, medical treatment, and adoption support, allowing users to quickly request help for injured or sick animals. The app also facilitates donations of food, medicine, and money, making it easy for users to contribute to the cause. Built with a user-friendly interface, Pawsitive ensures that every stray animal can receive timely care and attention, while fostering a community of compassionate animal lovers.